Create OrganizationAuthorization
This extracts out the code of authorizing against an organization so
that this can happen anywhere on or under the /organization/ORG path.
