+++ app/controllers/organizations_controller.rb
@@ -1,7 +1,10 @@
 class OrganizationsController < ApplicationController
-  before_action :authorize_organization_addition,                  only:   [:create]
-  before_action :set_organization, :authorize_organization_access, except: [:new, :create]
-  before_action :set_users_github_organizations,                   only:   [:new, :create]
+  include OrganizationAuthorization
+
+  before_action :authorize_organization_addition, only: [:create]
+  before_action :set_users_github_organizations,  only: [:new, :create]
+
+  skip_before_action :set_organization, :authorize_organization_access, only: [:new, :create]
 
   decorates_assigned :organization
 
@@ -59,15 +62,17 @@ class OrganizationsController < ApplicationController
     return if @organization.users.include?(current_user) || current_user.staff?
 
     begin
-      github_organization.admin?(github_user.login) ? @organization.users << current_user : not_found
+      github_organization.admin?(decorated_current_user.login) ? @organization.users << current_user : not_found
     rescue
       not_found
     end
   end
 
   def authorize_organization_addition
-    github_organization = GitHubOrganization.new(current_user.github_client, new_organization_params[:github_id].to_i)
-    deny_access unless github_organization.admin?(github_user.login)
+    new_github_organization = GitHubOrganization.new(current_user.github_client,
+                                                     new_organization_params[:github_id].to_i)
+
+    deny_access unless new_github_organization.admin?(decorated_current_user.login)
   end
 
   def deny_access
@@ -86,15 +91,13 @@ class OrganizationsController < ApplicationController
       .merge(users: [current_user])
   end
 
-  def not_found
-    fail ActionController::RoutingError, 'Not Found'
-  end
-
   def set_organization
     @organization = Organization.friendly.find(params[:id])
   end
 
   def set_users_github_organizations
+    github_user = GitHubUser.new(current_user.github_client)
+
     @users_github_organizations = github_user.admin_organization_memberships.map do |membership|
       unless Organization.find_by(github_id: membership.organization.id)
         [membership.organization.login, membership.organization.id]
@@ -106,20 +109,5 @@ class OrganizationsController < ApplicationController
     params
       .require(:organization)
       .permit(:title)
-  end
-
-  private
-
-  def github_organization
-    @existing_github_organization ||= GitHubOrganization.new(current_user.github_client, @organization.github_id)
-  end
-
-  def github_user
-    @github_user ||= GitHubUser.new(current_user.github_client)
-  end
-
-  def new_github_organization
-    @new_github_organization ||= GitHubOrganization.new(current_user.github_client,
-                                                        new_organization_params[:github_id].to_i)
   end
 end
