+++ app/controllers/users_controller.rb
@@ -1,6 +1,8 @@
 class UsersController < ApplicationController
   before_action :set_user
 
+  skip_before_action :set_organization, :authorize_organization_access
+
   def show
     @organizations = @user.organizations.page(params[:page])
   end
