+++ app/views/organizations/invite.html.erb
@@ -6,5 +6,6 @@
   </div>
 
   <div class="site-content-body">
+    <%= link_to 'skip for now', organization_path(@organization) %>
   </div>
 </div>
